More work on stuff

+++ inventory/InventoryItem.cs
@@ -5,11 +5,11 @@ namespace OOPAssignment011
         public static readonly string FoodType = "FOOD";
         public static readonly string ToyType = "TOY";
         public static readonly string MedicineType = "MEDICINE";
-        public float Cost { get; protected set; }
+        public int Cost { get; protected set; }
         public string Name { get; protected set; }
         public string Description { get; protected set; }
 
-        public InventoryItem(float cost, string name, string description)
+        public InventoryItem(int cost, string name, string description)
         {
             this.Cost = cost;
             this.Name = name;
